refactor(#4): Ajustando layouts para boards.
