+++ src/components/Dropdown.tsx
@@ -1,16 +1,19 @@
-import React, { useState } from "react";
+import React, { Fragment, useState } from "react";
+import { Menu, Transition } from "@headlessui/react";
 
 interface DropdownProps {
   label: string;
   items: string[];
-  icon?: string;
   onItemSelected: (item: string) => void;
+}
+
+function classNames(...classes: string[]): string {
+  return classes.filter(Boolean).join(" ");
 }
 
 const Dropdown: React.FC<DropdownProps> = ({
   label,
   items,
-  icon,
   onItemSelected,
 }) => {
   const [isOpen, setIsOpen] = useState(false);
@@ -20,49 +23,65 @@ const Dropdown: React.FC<DropdownProps> = ({
   };
 
   const handleItemClick = (item: string) => {
-    console.log(icon);
     onItemSelected(item);
     setIsOpen(false);
   };
 
   return (
-    <div className="dropdown">
-      <button
-        onClick={toggleDropdown}
-        id="actionsDropdown"
-        data-dropdown-toggle="actionsDropdown"
-        className="w-full md:w-auto flex items-center justify-center py-2 px-4 text-sm font-medium text-gray-900 focus:outline-none bg-white rounded-lg border border-gray-200 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700"
-        type="button"
-      >
-        {/* <svg
-          className="-ml-1 mr-1.5 w-5 h-5"
-          fill="currentColor"
-          viewBox="0 0 20 20"
-          xmlns="http://www.w3.org/2000/svg"
-          aria-hidden="true"
+    <Menu as="div" className="relative inline-block text-left">
+      <div>
+        <Menu.Button
+          onClick={toggleDropdown}
+          className="w-full md:w-auto flex items-center justify-center py-2 px-4 text-sm font-medium text-gray-900 focus:outline-none bg-white rounded-lg border border-gray-200 hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700"
         >
-          <path
-            clipRule="evenodd"
-            fillRule="evenodd"
-            d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
-          />
-        </svg> */}
+          <svg
+            className="-ml-1 mr-1.5 w-5 h-5"
+            fill="currentColor"
+            viewBox="0 0 20 20"
+            xmlns="http://www.w3.org/2000/svg"
+            aria-hidden="true"
+          >
+            <path
+              clipRule="evenodd"
+              fillRule="evenodd"
+              d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
+            />
+          </svg>
+          {label}
+        </Menu.Button>
+      </div>
 
-        {icon && <div dangerouslySetInnerHTML={{ __html: icon }} />}
-
-        {label}
-      </button>
-
-      {isOpen && (
-        <ul className="dropdown-menu">
-          {items.map((item, index) => (
-            <li key={index} onClick={() => handleItemClick(item)}>
-              {item}
-            </li>
-          ))}
-        </ul>
-      )}
-    </div>
+      <Transition
+        as={Fragment}
+        enter="transition ease-out duration-100"
+        enterFrom="transform opacity-0 scale-95"
+        enterTo="transform opacity-100 scale-100"
+        leave="transition ease-in duration-75"
+        leaveFrom="transform opacity-100 scale-100"
+        leaveTo="transform opacity-0 scale-95"
+      >
+        <Menu.Items className="absolute right-0 z-10 mt-2 w-56 origin-top-right rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+          <div className="py-1">
+            {items.map((item, index) => (
+              <Menu.Item key={index}>
+                {({ active }) => (
+                  <a
+                    href="#"
+                    onClick={() => handleItemClick(item)} // Call the function to handle item click
+                    className={classNames(
+                      active ? "bg-gray-100 text-gray-900" : "text-gray-700",
+                      "block px-4 py-2 text-sm"
+                    )}
+                  >
+                    {item}
+                  </a>
+                )}
+              </Menu.Item>
+            ))}
+          </div>
+        </Menu.Items>
+      </Transition>
+    </Menu>
   );
 };
